+++ packages/stargate/src/stargateclient.searchtx.spec.ts
@@ -190,12 +190,12 @@ describe("StargateClient.getTx and .searchTx", () => {
     });
   });
 
-  describe("with SearchByHeightQuery", () => {
+  describe("searchTx", () => {
     it("can search successful tx by height", async () => {
       pendingWithoutSimapp();
       assert(sendSuccessful, "value must be set in beforeAll()");
       const client = await StargateClient.connect(simapp.tendermintUrl);
-      const result = await client.searchTx({ height: sendSuccessful.height });
+      const result = await client.searchTx(`tx.height=${sendSuccessful.height}`);
       expect(result.length).toBeGreaterThanOrEqual(1);
       expect(result).toContain(
         jasmine.objectContaining({
@@ -211,7 +211,7 @@ describe("StargateClient.getTx and .searchTx", () => {
       pendingWithoutSimapp();
       assert(sendUnsuccessful, "value must be set in beforeAll()");
       const client = await StargateClient.connect(simapp.tendermintUrl);
-      const result = await client.searchTx({ height: sendUnsuccessful.height });
+      const result = await client.searchTx(`tx.height=${sendUnsuccessful.height}`);
       expect(result.length).toBeGreaterThanOrEqual(1);
       expect(result).toContain(
         jasmine.objectContaining({
@@ -222,14 +222,14 @@ describe("StargateClient.getTx and .searchTx", () => {
         }),
       );
     });
-  });
 
-  describe("with SearchBySentFromOrToQuery", () => {
     it("can search by sender", async () => {
       pendingWithoutSimapp();
       assert(sendSuccessful, "value must be set in beforeAll()");
       const client = await StargateClient.connect(simapp.tendermintUrl);
-      const results = await client.searchTx({ sentFromOrTo: sendSuccessful.sender });
+      // Since Cosmos SDK 0.47 we can only combine attributes of a single event
+      const query = `message.action = '/cosmos.bank.v1beta1.MsgSend' AND message.sender = '${sendSuccessful.sender}'`;
+      const results = await client.searchTx(query);
       expect(results.length).toBeGreaterThanOrEqual(1);
 
       // Check basic structure of all results
@@ -257,7 +257,7 @@ describe("StargateClient.getTx and .searchTx", () => {
       pendingWithoutSimapp();
       assert(sendSuccessful, "value must be set in beforeAll()");
       const client = await StargateClient.connect(simapp.tendermintUrl);
-      const results = await client.searchTx({ sentFromOrTo: sendSuccessful.recipient });
+      const results = await client.searchTx(`transfer.recipient='${sendSuccessful.recipient}'`);
       expect(results.length).toBeGreaterThanOrEqual(1);
 
       // Check basic structure of all results
@@ -281,69 +281,11 @@ describe("StargateClient.getTx and .searchTx", () => {
       );
     });
 
-    it("can search by recipient and filter by minHeight", async () => {
-      pendingWithoutSimapp();
-      assert(sendSuccessful);
-      const client = await StargateClient.connect(simapp.tendermintUrl);
-      const query = { sentFromOrTo: sendSuccessful.recipient };
-
-      {
-        const result = await client.searchTx(query, { minHeight: 0 });
-        expect(result.length).toEqual(1);
-      }
-
-      {
-        const result = await client.searchTx(query, { minHeight: sendSuccessful.height - 1 });
-        expect(result.length).toEqual(1);
-      }
-
-      {
-        const result = await client.searchTx(query, { minHeight: sendSuccessful.height });
-        expect(result.length).toEqual(1);
-      }
-
-      {
-        const result = await client.searchTx(query, { minHeight: sendSuccessful.height + 1 });
-        expect(result.length).toEqual(0);
-      }
-    });
-
-    it("can search by recipient and filter by maxHeight", async () => {
-      pendingWithoutSimapp();
-      assert(sendSuccessful);
-      const client = await StargateClient.connect(simapp.tendermintUrl);
-      const query = { sentFromOrTo: sendSuccessful.recipient };
-
-      {
-        const result = await client.searchTx(query, { maxHeight: 9999999999999 });
-        expect(result.length).toEqual(1);
-      }
-
-      {
-        const result = await client.searchTx(query, { maxHeight: sendSuccessful.height + 1 });
-        expect(result.length).toEqual(1);
-      }
-
-      {
-        const result = await client.searchTx(query, { maxHeight: sendSuccessful.height });
-        expect(result.length).toEqual(1);
-      }
-
-      {
-        const result = await client.searchTx(query, { maxHeight: sendSuccessful.height - 1 });
-        expect(result.length).toEqual(0);
-      }
-    });
-  });
-
-  describe("with SearchByTagsQuery", () => {
-    it("can search by transfer.recipient", async () => {
+    it("works with tags", async () => {
       pendingWithoutSimapp();
       assert(sendSuccessful, "value must be set in beforeAll()");
       const client = await StargateClient.connect(simapp.tendermintUrl);
-      const results = await client.searchTx({
-        tags: [{ key: "transfer.recipient", value: sendSuccessful.recipient }],
-      });
+      const results = await client.searchTx([{ key: "transfer.recipient", value: sendSuccessful.recipient }]);
       expect(results.length).toBeGreaterThanOrEqual(1);
 
       // Check basic structure of all results