+++ packages/stargate/src/testutils.spec.ts
@@ -23,35 +23,37 @@ export function simapp46Enabled(): boolean {
   return !!process.env.SIMAPP46_ENABLED;
 }
 
+export function simapp47Enabled(): boolean {
+  return !!process.env.SIMAPP47_ENABLED;
+}
+
 export function simappEnabled(): boolean {
-  return simapp44Enabled() || simapp46Enabled();
+  return simapp44Enabled() || simapp46Enabled() || simapp47Enabled();
 }
 
-export function pendingWithoutSimapp44Or46(): void {
-  if (!simapp44Enabled() && !simapp46Enabled()) {
-    return pending("Set SIMAPP{44,46}_ENABLED to enable Simapp based tests");
-  }
-}
-
-export function pendingWithoutSimapp46(): void {
-  if (!simapp46Enabled()) {
-    return pending("Set SIMAPP46_ENABLED to enable Simapp based tests");
+export function pendingWithoutSimapp46OrHigher(): void {
+  if (!simapp46Enabled() && !simapp47Enabled()) {
+    return pending("Set SIMAPP46_ENABLED or SIMAPP47_ENABLED to enable Simapp based tests");
   }
 }
 
 export function pendingWithoutSimapp(): void {
   if (!simappEnabled()) {
-    return pending("Set SIMAPP{44,46}_ENABLED to enable Simapp based tests");
+    return pending("Set SIMAPP{44,46,47}_ENABLED to enable Simapp based tests");
   }
 }
 
 export function slowSimappEnabled(): boolean {
-  return !!process.env.SLOW_SIMAPP44_ENABLED || !!process.env.SLOW_SIMAPP46_ENABLED;
+  return (
+    !!process.env.SLOW_SIMAPP44_ENABLED ||
+    !!process.env.SLOW_SIMAPP46_ENABLED ||
+    !!process.env.SLOW_SIMAPP47_ENABLED
+  );
 }
 
 export function pendingWithoutSlowSimapp(): void {
   if (!slowSimappEnabled()) {
-    return pending("Set SLOW_SIMAPP{44,46}_ENABLED to enable slow Simapp based tests");
+    return pending("Set SLOW_SIMAPP{44,46,47}_ENABLED to enable slow Simapp based tests");
   }
 }
 
@@ -155,7 +157,7 @@ export const validator = {
    */
   pubkey: {
     type: "tendermint/PubKeySecp256k1",
-    value: "AtDcuH4cX1eaxZrJ5shheLG3tXPAoV4awoIZmNQtQxmf",
+    value: "A0RZ3+xLf9xJiySHQxQsQtW8HJYEcniJKbFxG2R9ZEQv",
   },
   /**
    * delegator_address from /cosmos.staking.v1beta1.MsgCreateValidator in scripts/simapp44/template/.simapp/config/genesis.json
@@ -164,7 +166,7 @@ export const validator = {
    * jq ".app_state.genutil.gen_txs[0].body.messages[0].delegator_address" scripts/simapp44/template/.simapp/config/genesis.json
    * ```
    */
-  delegatorAddress: "cosmos1urk9gy7cfws0ak9x5nu7lx4un9n6gqkry79679",
+  delegatorAddress: "cosmos12nt2hqjps8r065wc02qks88tvqzdeua0ld3jxy",
   /**
    * validator_address from /cosmos.staking.v1beta1.MsgCreateValidator in scripts/simapp44/template/.simapp/config/genesis.json
    *
@@ -172,7 +174,7 @@ export const validator = {
    * jq ".app_state.genutil.gen_txs[0].body.messages[0].validator_address" scripts/simapp44/template/.simapp/config/genesis.json
    * ```
    */
-  validatorAddress: "cosmosvaloper1urk9gy7cfws0ak9x5nu7lx4un9n6gqkrp230jk",
+  validatorAddress: "cosmosvaloper12nt2hqjps8r065wc02qks88tvqzdeua06e982h",
   accountNumber: 0,
   sequence: 1,
 };